backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt-mongoengine==1.3.0
django-rest-framework-mongoengine==3.4.1
mongoengine==0.29.1
PyJWT==2.1.0
pymongo==4.12.0
python-dotenv==1.1.0
setuptools==78.1.0
drf-spectacular==0.28.0
django-cors-headers==3.7.0
djangorestframework-simplejwt==5.3.1
